% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinylight.R
\name{browseTo}
\alias{browseTo}
\title{Opens a browser to look at the server}
\usage{
browseTo(server)
}
\arguments{
\item{server}{The server to browse to}
}
\value{
No return value
}
\description{
Opens a browser to look at the server
}
