% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinylight.R
\name{encodePlotAs}
\alias{encodePlotAs}
\title{Renders a plot as a base64-encoded PNG}
\usage{
encodePlotAs(format, plotFn)
}
\arguments{
\item{format}{An object specifying the output, with the following members:
format$type is \code{"png"}, \code{"pdf"} or \code{"csv"}, and
\code{format$width} and \code{format$height} are
the dimensions of the PDF (in inches) or PNG (in pixels) if appropriate.}

\item{plotFn}{Function to call to perform the plot}
}
\value{
list with two keys, whose values can each be NULL:
\code{'plot'} is a plot in HTML img src form and \code{'data'} is a
data frame or other non-plot result.
}
\description{
The result can be set as the \code{src} attribute of an \code{<img>}
element in HTML.
}
\seealso{
\code{\link{rrpcServer}}
}
