% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinylight.R
\name{getInitResponse}
\alias{getInitResponse}
\title{Get the response to a POST to /init
This is index.html with (potentially) the JSON data
from the 'data' parameter inserted.}
\usage{
getInitResponse(req, path)
}
\arguments{
\item{req}{The httpuv request object}

\item{path}{File system path to the index.html file}
}
\value{
The httpuv response object
}
\description{
Get the response to a POST to /init
This is index.html with (potentially) the JSON data
from the 'data' parameter inserted.
}
