% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinylight.R
\name{indexWithInit}
\alias{indexWithInit}
\title{Get index.html with (potentially) the JSON data in `text`
inserted.}
\usage{
indexWithInit(text, path)
}
\arguments{
\item{text}{The text to insert as shinylight_initial_data}

\item{path}{File system path to the index.html file}
}
\value{
The updated text
}
\description{
Get index.html with (potentially) the JSON data in `text`
inserted.
}
