% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl.R
\name{rrpcServer}
\alias{rrpcServer}
\title{Makes and starts a server for serving R calculations}
\usage{
rrpcServer(
  interface,
  host = "0.0.0.0",
  port = NULL,
  appDirs = NULL,
  root = "/"
)
}
\arguments{
\item{interface}{List of functions to be served. The names of the elements
are the names that the client will use to call them.}

\item{host}{Interface to listen on (default is '0.0.0.0', that is, all interfaces)}

\item{port}{Port to listen on}

\item{appDirs}{List of directories in which to find static files to serve}

\item{root}{Root of the app on the server (with trailing slash)}
}
\value{
The server object, can be passed to [slStop]
}
\description{
Makes and starts a server for serving R calculations
}
