% Generated by JSDoc (with jsdoc-r template): do not edit by hand
% Please edit documentation in shinylight.js
\docType{package}
\name{shinylight.call}
\alias{shinylight.call}
\title{JavaScript function}
\section{Arguments}{\itemize{
\item{\code{fn} \code{string} The name of the R function to call.}
\item{\code{data} \code{object} An object whose keys are the arguments
to the function being called.}
\item{\code{plotElement} \code{string}, \code{HTMLElement} If provided, the
\code{<img>} element (or id of the element) that will receive the
plot output (if any). The plot returned will be the size that this
element already has, so ensure that it is styled in a way that it has
the correct size even if no image (or an old image) has been set.}
\item{\code{extra} \code{object} [optional] An object whose keys can be:
\code{"imgType"}: Type of image required, \code{"png"}
(default) or \code{"svg"}; \code{"info"}: Function to be called
if the R function \code{\link{sendInfoText}} is called;
\code{"progress"}: Function to be called if the R function
\code{\link{sendProgress}} is called.}
}}
\section{Return Value}{\itemize{
\item Result object that might have a \code{plot}
property (giving a string that would work as the \code{src}
attribute of an \code{img} element, representing graphics
drawn by the command), a \code{data} property (giving
the value returned by the command) and a \code{headers}
property (giving the column names in the data returned if any).
If the promise resolves to an error, the argument to the error
function is a string representing the cause of the error.
}}
\description{
Calls a server function as defined in the server's call to the
\code{slServer} function.
}