% Generated by JSDoc (with jsdoc-r template): do not edit by hand
% Please edit documentation in shinylight.js
\docType{package}
\name{shinylight.initialize}
\alias{shinylight.initialize}
\title{JavaScript function}
\description{
Call this before calling any other ShinyLight function.
Returns a promise that resolves (to nothing) when the
connection is ready.
}