% Generated by JSDoc (with jsdoc-r template): do not edit by hand
% Please edit documentation in shinylight.js
\docType{package}
\name{shinylight.setElementJson}
\alias{shinylight.setElementJson}
\title{JavaScript function}
\section{Arguments}{\itemize{
\item{\code{elementOrId} \code{string}, \code{HTMLElement} The element (or its id)
that will have its text set}
\item{\code{object} \code{any} The object whose JSON representation will be
set as the text content of the element}
}}
\description{
Sets the text content of an element (or its \code{value} as
appropriate) to the JSON representation of an object.
}