% Generated by JSDoc (with jsdoc-r template): do not edit by hand
% Please edit documentation in shinylight.js
\docType{package}
\name{shinylight.setGridResult}
\alias{shinylight.setGridResult}
\title{JavaScript function}
\section{Arguments}{\itemize{
\item{\code{grid} \code{DataEntryGrid} Table that receives the result}
\item{\code{result} \code{object} Return value promised by \code{\link{runR}}}
}}
\description{
Sets a \code{dataentrygrid} object to the result of
\code{\link{runR}}, if appropriate.
}