% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinylight.R
\name{slRunRServer}
\alias{slRunRServer}
\title{Start a ShinyLight server which runs R that it is sent}
\usage{
slRunRServer(
  permittedSymbols,
  appDir = NULL,
  host = "0.0.0.0",
  port = NULL,
  daemonize = FALSE
)
}
\arguments{
\item{permittedSymbols}{List of symbols that are permitted in the R
commands passed. Remember to include \code{data}, \code{$} and
\code{<-}.}

\item{appDir}{Directory containing files to serve (for example
system.file("www", package = "your-package"))}

\item{host}{IP address to listen on, default is 0.0.0.0 (all interfaces)}

\item{port}{Internet port of the virtual server. If not defined, a
random free port will be chosen and the browser will be opened
to show the GUI.}

\item{daemonize}{If TRUE, keep serving forever without returning.
This is useful when called from RScript, to keep}
}
\value{
server object, unless daemonize is TRUE.
}
\description{
Start a ShinyLight server which runs R that it is sent
}
