% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl.R
\name{slServer}
\alias{slServer}
\title{Start a ShinyLight server}
\usage{
slServer(appDir, interface, host = "0.0.0.0", port = NULL, daemonize = FALSE)
}
\arguments{
\item{appDir}{Directory containing files to serve (for example
system.file("www", package = "your-package"))}

\item{interface}{List of functions you want to be able to call from
the browser}

\item{host}{IP address to listen on, default is 0.0.0.0 (all interfaces)}

\item{port}{Internet port of the virtual server. If not defined, a
random free port will be chosen and the browser will be opened
to show the GUI.}

\item{daemonize}{If TRUE, keep serving forever without returning.
This is useful when called from RScript, to keep}
}
\value{
server object, unless daemonize is TRUE.
}
\description{
Start a ShinyLight server
}
