% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinylight.R
\name{slStop}
\alias{slStop}
\title{Stops a ShinyLight GUI}
\usage{
slStop(server = NULL)
}
\arguments{
\item{server}{The server (returned by \code{\link{slServer}}
or \code{\link{slRunRServer}})
to stop. If not supplied all servers will be stopped.}
}
\description{
Stops a ShinyLight GUI
}
