% Generated by JSDoc (with jsdoc-r template): do not edit by hand
% Please edit documentation in shinylight-toolkit.js
\docType{package}
\name{toolkit.banner}
\alias{toolkit.banner}
\title{JavaScript function}
\section{Arguments}{\itemize{
\item{\code{elements} \code{Array.<HTMLControlElement>} Initial array of elements to be added.}
\item{\code{className} \code{string} HTML class for the returned banner.}
}}
\section{Return Value}{\itemize{
\item The banner element.
}}
\description{
Returns a Container Element for displaying controls horizontally.
}