% Generated by JSDoc (with jsdoc-r template): do not edit by hand
% Please edit documentation in shinylight-toolkit.js
\docType{package}
\name{toolkit.footer}
\alias{toolkit.footer}
\title{JavaScript function: A panel with a smaller footer.}
\section{Arguments}{\itemize{
\item{\code{ftr} \code{HTMLElement} The footer element.}
\item{\code{main} \code{\link{toolkit.HTMLPositionedElement}} The body element.}
}}
\section{Return Value}{\itemize{
\item The element containing
the footer and body.
}}
\description{
Returns a Positioned Element consisting of a body and a footer.
}