% Generated by JSDoc (with jsdoc-r template): do not edit by hand
% Please edit documentation in shinylight-toolkit.js
\docType{package}
\name{toolkit.forEach}
\alias{toolkit.forEach}
\title{JavaScript function}
\section{Arguments}{\itemize{
\item{\code{a} \code{object} Object or array to be iterated through.}
\item{\code{f} \code{function} Function to call with two arguments: the key
of the element (or index in the case of an array) and the value.}
}}
\description{
Calls a function for each member of an array or object.
}