% Generated by JSDoc (with jsdoc-r template): do not edit by hand
% Please edit documentation in shinylight-toolkit.js
\docType{package}
\name{toolkit.header}
\alias{toolkit.header}
\title{JavaScript function: A panel with a smaller header.}
\section{Arguments}{\itemize{
\item{\code{hdr} \code{HTMLElement} The header element.}
\item{\code{main} \code{\link{toolkit.HTMLPositionedElement}} The body element.}
}}
\section{Return Value}{\itemize{
\item The element containing
the header and body.
}}
\description{
Returns a Positioned Element consisting of a header and a body.
}