% Generated by JSDoc (with jsdoc-r template): do not edit by hand
% Please edit documentation in shinylight-toolkit.js
\name{toolkit.HTMLContainerElement}
\alias{toolkit.HTMLContainerElement}
\title{JavaScript class: A monkey-patched \code{HTMLElement}.}
\description{
A Container Element is an element for displaying a set of controls and their labels.
}
\section{Properties}{\describe{
\item{makeSubElement}{\code{function} Gets an element in which a control and its label can be stored. You do not need
to call this unless you have made your own custom control; it will be called
by functions such as \code{\link{toolkit.paramText}}. Pass in the ID of the control
(you will need the ID for the \code{getData} and \code{setData} calls).}
\item{getData}{\code{function} Returns an object mapping contained
controls (or nested containers) to their current values.}
\item{setData}{\code{function} Sets the values
of the contained controls. \code{data} is a mapping from the IDs of
the contained controls to the data that should be set on them.}
}}
\seealso{
toolkit.stack

toolkit.banner

toolkit.optionsPage
}