% Generated by JSDoc (with jsdoc-r template): do not edit by hand
% Please edit documentation in shinylight-toolkit.js
\name{toolkit.HTMLControlContainerElement}
\alias{toolkit.HTMLControlContainerElement}
\title{JavaScript class}
\description{
A container for a single control.
}
\section{Properties}{\describe{
\item{addElement}{\code{function} Adds an element. Should be called once with a control's label, and then
again with the control itself.}
}}
\seealso{
\code{\link{toolkit.HTMLContainerElement}}
}