% Generated by JSDoc (with jsdoc-r template): do not edit by hand
% Please edit documentation in shinylight-toolkit.js
\docType{package}
\name{toolkit.HTMLControlElement}
\alias{toolkit.HTMLControlElement}
\title{JavaScript class}
\description{
A monkey-patched \code{HTMLElement} representing a control with its
label.
}
\section{Properties}{\describe{
\item{getData}{\code{function} Returns the current displayed value.}
\item{setData}{\code{function} Sets the value.}
\item{hide}{\code{function} Makes the element invisible and non-interactive}
\item{show}{\code{function} makes the element visible and (potentially) interactive}
}}
\seealso{
\code{\link{toolkit.paramBoolean}}

\code{\link{toolkit.paramColor}}

\code{\link{toolkit.paramFloat}}

\code{\link{toolkit.paramInteger}}

\code{\link{toolkit.paramSelector}}

\code{\link{toolkit.paramText}}
}