% Generated by JSDoc (with jsdoc-r template): do not edit by hand
% Please edit documentation in shinylight-toolkit.js
\docType{package}
\name{toolkit.HTMLPositionedElement}
\alias{toolkit.HTMLPositionedElement}
\title{JavaScript class: A monkey-patched \code{HTMLElement} with some extra methods.}
\description{
Certain elements returned by Toolkit methods are Positioned Elements. It is
necessary for elements in some places in the document to be Positioned
Elements for the document resizing and formatting to work.

If you have an HTML element that is not a Positioned Element that you want
to add to a place where only Positioned Elements are required, wrap it in
\code{\link{toolkit.scrollingWrapper}} or \code{\link{toolkit.nonScrollingWrapper}}.
}
\section{Properties}{\describe{
\item{setSize}{\code{function} Sets the position of the element on the document in pixels, with parameters
for left, top, width and height in that order.}
\item{getSize}{\code{function} Returns an object with members \code{left}, \code{top}, \code{width}
and \code{height} for the position of the element.}
\item{hide}{\code{function} Makes the element invisible and non-interactive}
\item{show}{\code{function} makes the element visible and (potentially) interactive}
}}
