% Generated by JSDoc (with jsdoc-r template): do not edit by hand
% Please edit documentation in shinylight-toolkit.js
\docType{package}
\name{toolkit.image}
\alias{toolkit.image}
\title{JavaScript function}
\section{Arguments}{\itemize{
\item{\code{updateSizeFunction} \code{function} Nullary function called when
the object's size is changed.}
}}
\section{Return Value}{\itemize{
\item Image element. It has a
\code{getSize()} method, returning an object with width and height
members. This is the width and height set by \code{reposition()},
not the actual on-screen width and height, if that is different
for some reason. In other words, it returns the width and height
the image "should" have.
}}
\description{
An image element.
}