% Generated by JSDoc (with jsdoc-r template): do not edit by hand
% Please edit documentation in shinylight-toolkit.js
\docType{package}
\name{toolkit.nonScrollingWrapper}
\alias{toolkit.nonScrollingWrapper}
\title{JavaScript function: Returns a Positioned Element just containing one element.}
\section{Arguments}{\itemize{
\item{\code{element} \code{HTMLElement} The element to be wrapped}
\item{\code{verticalPadding} \code{int} The number of extra pixels above the element's
height to use as the returned element's default height.}
\item{\code{horizontalPadding} \code{int} The number of extra pixels above the element's
width to use as the returned element's default width.}
}}
\section{Return Value}{\itemize{
\item The wrapper.
}}
\description{
This element does not gain scrollbars if it is too large for this returned
container, and it will try to take up its full size in the layout.
}