% Generated by JSDoc (with jsdoc-r template): do not edit by hand
% Please edit documentation in shinylight-toolkit.js
\docType{package}
\name{toolkit.optionsPage}
\alias{toolkit.optionsPage}
\title{JavaScript function: Returns a Container Element for displaying controls vertically.}
\section{Return Value}{\itemize{
\item A Container Element for displaying elements vertically.
}}
\description{
Returns an element with a \code{makeSubElement} method that
adds elements vertically. This differs from \code{\link{toolkit.stack}} in that the
labels will be aligned on the left and the controls will be aligned on the
right. It would make a nice options page, for example.
}