% Generated by JSDoc (with jsdoc-r template): do not edit by hand
% Please edit documentation in shinylight-toolkit.js
\docType{package}
\name{toolkit.paramFloat}
\alias{toolkit.paramFloat}
\title{JavaScript function: Returns a floating point input Toolkit Control.}
\section{Arguments}{\itemize{
\item{\code{id} \code{string} when \code{getData} or \code{setData} is
called on the container, the value at \code{'id'} refers to this
selector. The HTML id is set to \code{'param-' + id}.}
\item{\code{container} \code{\link{toolkit.HTMLContainerElement}} [optional] Where to put the control.}
\item{\code{translations} \code{object} Optional mapping: \code{translations.id}
is the name of the control to be displayed and \code{translations.help}
is help text to be displayed if the user hovers over the label}
\item{\code{initial} \code{string} Optional initial value for the control}
\item{\code{callback} \code{function} Optional function to be called whenever the
input value changes}
\item{\code{min} \code{float} Minimum permitted value (optional).}
\item{\code{max} \code{float} Maximum permitted value (optional).}
}}
\section{Return Value}{\itemize{
\item Text input control.
}}
\description{
Values outside the permitted range will gain the "invalid" class,
but there is no other effect.
}