% Generated by JSDoc (with jsdoc-r template): do not edit by hand
% Please edit documentation in shinylight-toolkit.js
\docType{package}
\name{toolkit.paramSelector}
\alias{toolkit.paramSelector}
\title{JavaScript function: Returns a custom selection box Toolkit Control.}
\section{Arguments}{\itemize{
\item{\code{id} \code{string} when \code{getData} or \code{setData} is
called on the container, the value at \code{'id'} refers to this
selector. The HTML id is set to \code{'param-' + id}.}
\item{\code{container} \code{\link{toolkit.HTMLContainerElement}} [optional] Where to put the control.
the container came from \code{optionsPage()} the new selection box
will be formatted as a table row.}
\item{\code{labelTranslations} \code{object} A dictionary with two optional keys;
'name' gives the label to display and 'help' gives HTML help text.
'help' has no effect unless 'name' is also present.}
\item{\code{values} \code{Array.<int>} An array of the IDs of the options
in the selection.}
\item{\code{valueTranslations} \code{object} A dictionary whose keys are the
IDs of the options in the selection, the values are more dictionaries.
These dictionaries have two optional keys; \code{'name'} (giving
the name to display for this option) and \code{'help'} (giving tooltip
HTML text).}
\item{\code{initial} \code{string} ID of the option to start selecting (optional)}
\item{\code{callback} \code{function} The (nullary) function to call when the
value changes (optional)}
}}
\section{Return Value}{\itemize{
\item The selection box.
}}
\description{
This is different to a normal selection box because it allows
tooltips on the items within the list.
}