% Generated by JSDoc (with jsdoc-r template): do not edit by hand
% Please edit documentation in shinylight-toolkit.js
\docType{package}
\name{toolkit.progressBar}
\alias{toolkit.progressBar}
\title{JavaScript function: Returns a Positioned Element progress bar.}
\section{Return Value}{\itemize{
\item The progress bar element.
}}
\description{
The progress is set by calling the \code{setData()} method.
}