% Generated by JSDoc (with jsdoc-r template): do not edit by hand
% Please edit documentation in shinylight-toolkit.js
\docType{package}
\name{toolkit.rightSideBar}
\alias{toolkit.rightSideBar}
\title{JavaScript function: A panel with a side bar.}
\section{Arguments}{\itemize{
\item{\code{bar} \code{HTMLElement} The side bar element.}
\item{\code{main} \code{\link{toolkit.HTMLPositionedElement}} The body element.}
}}
\section{Return Value}{\itemize{
\item The Toolkit Positioned Element containing
the side bar and body.
}}
\description{
Returns a Positioned Element consisting of a right side bar and a body.
}