% Generated by JSDoc (with jsdoc-r template): do not edit by hand
% Please edit documentation in shinylight-toolkit.js
\docType{package}
\name{toolkit.setAsBody}
\alias{toolkit.setAsBody}
\title{JavaScript function: Replaces the \code{<main>} tag in the document with this element.}
\section{Arguments}{\itemize{
\item{\code{el} \code{HTMLElement} The element to set as \code{<main>}}
}}
\description{
The element will have its \code{resize} event wired up. If \code{el}
is a Toolkit Positioned Element, it will be resized correctly when the
window is resized.
}