% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinylight.R
\name{downloadCsv}
\alias{downloadCsv}
\title{Encodes a data frame as a CSV file to be downloaded}
\usage{
downloadCsv(results)
}
\arguments{
\item{results}{Data frame to be returned}
}
\value{
A list to be returned to the browser describing a CSV
file to be downloaded.
}
\description{
Encodes a data frame as a CSV file to be downloaded
}
