% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl.R
\name{encodePlotAs}
\alias{encodePlotAs}
\title{Renders a plot as a base64-encoded PNG}
\usage{
encodePlotAs(format, plotFn)
}
\arguments{
\item{format}{An object specifying the output, with the following members:
format$type is "png", "pdf" or "csv", and format$width and format$height are
the dimensions of the PDF (in inches) or PNG (in pixels) if appropriate.}

\item{plotFn}{Function to call to perform the plot}
}
\value{
list with two keys, whose values can each be NULL:
'plot' is a plot in HTML img src form and 'data' is a data frame or other
non-plot result.
}
\description{
The result can be set as the `src` attribute of an `img` element in HTML.
}
\seealso{
[rrpcServer()]
}
