% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinylight.R
\name{getAddress}
\alias{getAddress}
\title{Obtains the address that the server is listening on}
\usage{
getAddress(server)
}
\value{
protocol://address:port
}
\description{
Obtains the address that the server is listening on
}
