% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl.R
\name{runR}
\alias{runR}
\title{Returns a function that runs an R command}
\usage{
runR(symbolList)
}
\arguments{
\item{symbolList}{A list of permitted symbols in the R command}
}
\description{
Returns a function that runs an R command
}
