% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinylight.R
\name{sendProgress}
\alias{sendProgress}
\title{Sends a progress update to the client.}
\usage{
sendProgress(numerator, denominator = 1)
}
\arguments{
\item{numerator}{The progress, out of [denominator]}

\item{denominator}{What the progress is out of. You could use this for the
number of known items to be completed so that each call increases either
the numerator (for more items done) and/or the denominator (for more items
discovered that need to be done). However, it is not necessary to do this, you
can reduce the numerator if you want.}
}
\description{
During a slow remote procedure call, call this to inform the client of progress.
}
