% Generated by JSDoc (with jsdoc-r template): do not edit by hand
% Please edit documentation in shinylight.js
\name{shinylight.call}
\alias{shinylight.call}
\title{JavaScript function}
\arguments{
\item{fn}{\code{string} The name of the R function to call.}
\item{data}{\code{object} An object whose keys are the arguments
to the function being called.}
\item{plotElement}{\code{string}, \code{HTMLElement} If provided, the
\code{<img>} element (or id of the element) that will receive the
plot output (if any). The plot returned will be the size that this
element already has, so ensure that it is styled in a way that it has
the correct size even if no image (or an old image) has been set.}
\item{extra}{\code{object} [optional] An object whose keys can be:
"imgType": Type of image required, "png" (default) or "svg";
"info": Funtion to be called if the R function \code{\link{sendInfoText}}
is called; "progress": Function to be called if the R function
\code{\link{sendProgress}} is called.}
}
\description{
Calls a server function as defined in the server's call to the
\code{slServer} function.
}