% Generated by JSDoc (with jsdoc-r template): do not edit by hand
% Please edit documentation in shinylight.js
\docType{package}
\name{shinylight.makeTable}
\alias{shinylight.makeTable}
\title{JavaScript function}
\arguments{
\item{data}{\code{object} Data as returned from R}
\item{extraColumns}{\code{Array.<string>}, \code{number} The extra column headers
required or the number of extra columns required.}
}
\value{
Headers and rows
}
\description{
Turns data received from R into a form that can be set into
dataentrygrid.js.
}
\examples{\dontrun{
t = shinylight.makeTable(data);
grid.init(t.headers, t.rows);
}}
