% Generated by JSDoc (with jsdoc-r template): do not edit by hand
% Please edit documentation in shinylight.js
\docType{package}
\name{shinylight.runR}
\alias{shinylight.runR}
\title{JavaScript function: Runs an R function.}
\arguments{
\item{rCommand}{\code{string} The R text to run. It can plot a graph
and/or return some R data structure (such as a data frame).}
\item{data}{\code{any} A javascript value that will be translated
to the R command as a value also called 'data'.}
\item{plotElement}{\code{string}, \code{HTMLElement} If provided, the
\code{<img>} element (or id of the element) that will receive the
plot output (if any). The plot returned will be the size that this
element already has, so ensure that it is styled in a way that it has
the correct size even if no image (or an old image) has been set.}
\item{extra}{\code{object} [optional] An object whose keys can be:
\code{"imgType"}: Type of image required, \code{"png"}
(default) or \code{"svg"}; \code{"info"}: Function to be called
if the R function \code{\link{sendInfoText}} is called;
\code{"progress"}: Function to be called if the R function
\code{\link{sendProgress}} is called.}
}
\value{
Result object that might have a \code{plot}
property (giving a string that would work as the \code{src}
attribute of an \code{img} element, representing graphics
drawn by the command) and a \code{data} property (giving
the value returned by the command). If the promise resolves
to an error, the argument to the error function is a string
representing the cause of the error.
}
\description{
The R side must be running the slRunRServer function.
}