% Generated by JSDoc (with jsdoc-r template): do not edit by hand
% Please edit documentation in shinylight.js
\name{shinylight.setElementPlot}
\alias{shinylight.setElementPlot}
\title{JavaScript function: Sets an \code{<img>} element to display a plot returned by
\code{\link{runR}}.}
\arguments{
\item{elementOrId}{\code{string}, \code{HTMLImageElement} The
\code{<img>} element (or its id) that will receive the image.}
\item{result}{\code{object} The result from \code{\link{runR}}.}
}
\description{
Normally you do not need to call this because to get
\code{shinylight} to produce a plot you need to set the
\code{plotElement} argument, and doing so will cause this
element to receive the plot automatically.
}