% Generated by JSDoc (with jsdoc-r template): do not edit by hand
% Please edit documentation in shinylight.js
\docType{package}
\name{shinylight.setElementText}
\alias{shinylight.setElementText}
\title{JavaScript function}
\arguments{
\item{elementOrId}{\code{string}, \code{HTMLElement} The element (or its id)
that will have its text set}
\item{text}{\code{string} The text to set into the element}
}
\description{
Sets the text content of an element (or its \code{value} as
appropriate).
}