% Generated by JSDoc (with jsdoc-r template): do not edit by hand
% Please edit documentation in shinylight.js
\name{shinylight.setGridResultWithNamedRows}
\alias{shinylight.setGridResultWithNamedRows}
\title{JavaScript function}
\arguments{
\item{grid}{\code{DataEntryGrid} Table that receives the result}
\item{result}{\code{object} Return value promised by \code{\link{runR}}}
}
\description{
Sets a \code{dataentrygrid} object to the result of
\code{\link{runR}}. The object will have fixed rows, with
names derived from the row names in the original
data frame.
}