% Generated by JSDoc (with jsdoc-r template): do not edit by hand
% Please edit documentation in shinylight-toolkit.js
\name{toolkit.banner}
\alias{toolkit.banner}
\title{JavaScript function}
\arguments{
\item{elements}{\code{Array.<HTMLControlElement>} Initial array of elements to be added.}
\item{className}{\code{string} HTML class for the returned banner.}
}
\description{
Returns a Container Element for displaying controls horizontally.
}