% Generated by JSDoc (with jsdoc-r template): do not edit by hand
% Please edit documentation in shinylight-toolkit.js
\docType{package}
\name{toolkit.deref}
\alias{toolkit.deref}
\title{JavaScript function: Dereferences an object or array through multiple indices.}
\arguments{
\item{object}{\code{object} The object to be dereferenced.}
\item{path}{\code{Array} The series of indices to be applied.}
\item{defaultValue}{\code{\link{toolkit.any}} The default value to be returned if
the path cannot be followed to the end.}
}
\description{
\code{deref(o, [a,b,c], d)} is a safe way of doing
\code{o[a][b][c]}. If that path does not exist, d is returned.
If d is not supplied, null is returned. Any undefined values in
path are ignored.
}