% Generated by JSDoc (with jsdoc-r template): do not edit by hand
% Please edit documentation in shinylight-toolkit.js
\name{toolkit.footer}
\alias{toolkit.footer}
\title{JavaScript function}
\arguments{
\item{ftr}{\code{HTMLElement} The footer element.}
\item{main}{\code{\link{toolkit.HTMLPositionedElement}} The body element.}
}
\description{
Returns a Positioned Element consisting of a body and a footer.
}