% Generated by JSDoc (with jsdoc-r template): do not edit by hand
% Please edit documentation in shinylight-toolkit.js
\name{toolkit.groupTitle}
\alias{toolkit.groupTitle}
\title{JavaScript function}
\arguments{
\item{container}{\code{HTMLElement} The container, preferably the
return value from \code{\link{toolkit.optionsPage}}.}
\item{labelTranslations}{\code{object} An object with two keys:
\code{'name'} is the display text for this title,
\code{'help'} (optional) is the tooltip text.}
}
\description{
Adds a group title to an \code{\link{toolkit.optionsPage}}.
}