% Generated by JSDoc (with jsdoc-r template): do not edit by hand
% Please edit documentation in shinylight-toolkit.js
\name{toolkit.header}
\alias{toolkit.header}
\title{JavaScript function}
\arguments{
\item{hdr}{\code{HTMLElement} The header element.}
\item{main}{\code{\link{toolkit.HTMLPositionedElement}} The body element.}
}
\description{
Returns a Positioned Element consisting of a header and a body.
}