% Generated by JSDoc (with jsdoc-r template): do not edit by hand
% Please edit documentation in shinylight-toolkit.js
\docType{package}
\name{toolkit.image}
\alias{toolkit.image}
\title{JavaScript function}
\arguments{
\item{updateSizeFunction}{\code{function} Nullary function called when
the object's size is changed.}
}
\description{
An image element.
}