% Generated by JSDoc (with jsdoc-r template): do not edit by hand
% Please edit documentation in shinylight-toolkit.js
\docType{package}
\name{toolkit.leftSideBar}
\alias{toolkit.leftSideBar}
\title{JavaScript function: A panel with a side bar.}
\arguments{
\item{bar}{\code{HTMLElement} The side bar element.}
\item{main}{\code{\link{toolkit.HTMLPositionedElement}} The body element.}
}
\description{
Returns a Positioned Element consisting of a left side bar and a body.
}