% Generated by JSDoc (with jsdoc-r template): do not edit by hand
% Please edit documentation in shinylight-toolkit.js
\name{toolkit.makeLabel}
\alias{toolkit.makeLabel}
\title{JavaScript function}
\arguments{
\item{translations}{\code{object} \code{translations[id].name} is the string
to use as label's text, \code{translations[id].help} is the string to use as
the label's tooltip. If \code{id} is undefined or null, \code{translations.name}
and \code{translations.help} are used.}
\item{container}{\code{\link{toolkit.HTMLControlContainerElement}} [optional] Where to put the
label.}
\item{id}{\code{string} [optional] Where to look in \code{translations} for the text.}
\item{idFor}{\code{string} [optional] The \code{id} attribute of the HTML element
that this element refers to.}
}
\description{
Makes a label suitable for labelling a control.
}