% Generated by JSDoc (with jsdoc-r template): do not edit by hand
% Please edit documentation in shinylight-toolkit.js
\docType{package}
\name{toolkit.overlay}
\alias{toolkit.overlay}
\title{JavaScript function: A panel with an overlay.}
\arguments{
\item{overlay}{\code{HTMLElement} The higher element. Any
\code{getData()} or \code{setData()} call on the
returned element will not be passed on to this overlay element.}
\item{main}{\code{\link{toolkit.HTMLPositionedElement}} The lower element.}
}
\value{
The element containing both elements.
}
\description{
Returns a Positioned Element consisting of two elements
placed in the same position. To be able to see the lower (main)
element you must either call \code{hide()} on the overlay,
or make it transparent with CSS.
}