% Generated by JSDoc (with jsdoc-r template): do not edit by hand
% Please edit documentation in shinylight-toolkit.js
\name{toolkit.progressBar}
\alias{toolkit.progressBar}
\title{JavaScript function: Returns a Positioned Element progress bar.}
\description{
The progress is set by calling the \code{setData()} method.
}