% Generated by JSDoc (with jsdoc-r template): do not edit by hand
% Please edit documentation in shinylight-toolkit.js
\name{toolkit.rightSideBar}
\alias{toolkit.rightSideBar}
\title{JavaScript function}
\arguments{
\item{bar}{\code{HTMLElement} The side bar element.}
\item{main}{\code{\link{toolkit.HTMLPositionedElement}} The body element.}
}
\description{
Returns a Positioned Element consisting of a right side bar and a body.
}