% Generated by JSDoc (with jsdoc-r template): do not edit by hand
% Please edit documentation in shinylight-toolkit.js
\docType{package}
\name{toolkit.scrollingWrapper}
\alias{toolkit.scrollingWrapper}
\title{JavaScript function: Returns a Positioned Element just containing one element.}
\arguments{
\item{element}{\code{HTMLElement} The element to be wrapped}
\item{verticalPadding}{\code{int} The number of extra pixels above the element's
height to use as the returned element's default height.}
\item{horizontalPadding}{\code{int} The number of extra pixels above the element's
width to use as the returned element's default width.}
}
\value{
The wrapper.
}
\description{
This element gains scrollbars if it is too large for this returned container.
}