% Generated by JSDoc (with jsdoc-r template): do not edit by hand
% Please edit documentation in shinylight-toolkit.js
\name{toolkit.stack}
\alias{toolkit.stack}
\title{JavaScript function: Returns a Container Element for displaying controls
vertically.}
\arguments{
\item{elements}{\code{Array.<HTMLElement>} Initial array of elements to be added.}
}
\description{
Returns a Container Element with a
\code{makeSubElement} method that adds elements vertically,
with the labels above the controls they correspond to.
}