% Generated by JSDoc (with jsdoc-r template): do not edit by hand
% Please edit documentation in shinylight-toolkit.js
\name{toolkit.staticText}
\alias{toolkit.staticText}
\title{JavaScript function: A static text Toolkit Control.}
\arguments{
\item{id}{\code{string} The ID of this control within the container}
\item{container}{\code{\link{toolkit.HTMLContainerElement}} [optional] Where to put the control.}
\item{translations}{\code{object} An object with keys \code{'name'}
for the label displayed by the text and \code{'help'} for tooltop text.}
}
\description{
This element is like a control in that it has a label and actual
text content, but it is not interactive.
}