% Generated by JSDoc (with jsdoc-r template): do not edit by hand
% Please edit documentation in shinylight-toolkit.js
\name{toolkit.verticalDivide}
\alias{toolkit.verticalDivide}
\title{JavaScript function}
\arguments{
\item{container}{\code{\link{toolkit.HTMLPositionedElement}} The container to divide.
If null, a container will be created for you.}
\item{left}{\code{\link{toolkit.HTMLPositionedElement}} The element to put on
the left of the divider.}
\item{right}{\code{\link{toolkit.HTMLPositionedElement}} The element to put
on the right of the divider.}
}
\description{
Returns a Positioned Element with a draggable vertical divider
bordering two other Positioned Elements.
}