% Generated by JSDoc (with jsdoc-r template): do not edit by hand
% Please edit documentation in shinylight-toolkit.js
\docType{package}
\name{toolkit.withTimeout}
\alias{toolkit.withTimeout}
\title{JavaScript function: Adds a fake callback argument to a nullary function.}
\arguments{
\item{fn}{\code{function} Nullary function to wrap.}
}
\description{
Perhaps you have a nullary function that you want called
when the user clicks a button, but the \code{\link{toolkit.button}}
function wants a unary function that has a completion callback
so that the button knows when to pop back up again. In this
situation you might wrap your function with a call to
\code{\link{toolkit.withTimeout}}.
}